+++ sitewit/services.py
@@ -7,6 +7,10 @@ from yoconfig import get_config
 import sitewit
 
 
+def _remove_nones(data):
+    return {k: v for k, v in data.items() if v is not None}
+
+
 class SitewitService(HTTPServiceClient):
     """Client for SiteWit's API.
 
@@ -71,7 +75,7 @@ class SitewitService(HTTPServiceClient):
             user_email (str): email of account owner.
             currency (str): user's currency.
             country_code (str): user's location.
-            site_id (str): site ID (uuid4).
+            site_id (str, optional): site ID (uuid4).
             user_token (str, optional): user token in case this account is
                 owned by existing user.
             remote_subpartner_id (str, optional): user's partner_id on
@@ -85,7 +89,6 @@ class SitewitService(HTTPServiceClient):
             'url': url,
             'businessType': 'SMB',
             'timeZone': self.DEFAULT_TIME_ZONE,
-            'clientId': site_id,
             'name': user_name,
             'email': user_email,
             'currency': currency,
@@ -93,6 +96,9 @@ class SitewitService(HTTPServiceClient):
         }
         if user_token is not None:
             data['userToken'] = user_token
+
+        if site_id:
+            data['clientId'] = site_id
 
         return self.post(
             '/api/account/', json=data,
@@ -113,7 +119,8 @@ class SitewitService(HTTPServiceClient):
             '/api/account/',
             headers=self._get_account_auth_header(account_token)).json()
 
-    def update_account(self, account_token, url, country_code, currency):
+    def update_account(
+            self, account_token, url=None, country_code=None, currency=None):
         """Update SiteWit account.
 
         Args:
@@ -123,11 +130,12 @@ class SitewitService(HTTPServiceClient):
             account json:
             https://sandboxpapi.sitewit.com/Help/Api/GET-api-Account
         """
-        data = {
+        data = _remove_nones({
             'url': url,
             'countryCode': country_code,
             'currency': currency,
-        }
+        })
+
 
         return self.put(
             '/api/account/', json=data,