+++ tests/accounts/test_integration.py
@@ -152,8 +152,8 @@ class TestUpdateAccount(AccountTestCase):
         self.created_account = self.create_account()
 
         self.updated_account = service.update_account(
-            self.created_account['accountInfo']['token'], 'http://url.new',
-            'GB', 'GBP')
+            self.created_account['accountInfo']['token'],
+            url='http://url.new', country_code='GB', currency='GBP')
 
         self.retrieved_account = service.get_account(
             self.updated_account['token'])