+++ CHANGELOG.md
@@ -2,9 +2,11 @@
 
 ## DEV
 
+* Added `models.Account.set_site_id()` method.
+* Added `.Account.set_site_id()` method.
 * Re-ordered parameters of `models.Account.create()` and
-  `services.SitewitService.create()`. `site_id/client_id` parameter made
-  optional.
+  `services.SitewitService.create()`.
+* `site_id/client_id` parameter made optional for account creation.
 
 ## 0.7.3